requirements.txt
langchain
langchain-core
langchain-community

langchain-openai
openai

langchain-anthropic

langchain-google-genai
google-generativeai

langchain-huggingface
transformers
huggingface-hub

python-dotenv

numpy
scikit-learn

pypdf
bs4